build-system: chromium-binary-build: Extend wrapper-plan syntax.
This commit allows patchelf-plan to take entries with an optional path.  Here,
wrapper-plan is extended to allow for additional syntax (not just a list of
strings) similar to patchelf-plan.  Now, entries can be a list, with the first
the string for the file to be patched and the second a list which is added to
the patchelf-plan.

This allows, for example, to patch RPATH to effectively have $ORIGIN for
binaries that need it, with an entry like `("bin/binary" (("out"
"/lib/Binary")))` common for some chromium-based packages.