packages: firefox: Fix video decoding/acceleration.
As of Firefox 137, VA-API video acceleration is enabled by default.  This
would fail due to not finding libpciaccess.  This is an indirect dependency,
through libdrm (that mesa and libva depend on).  It would be best to have our
runpaths-of-input work recursively to catch this (so it can be in the RDD
sandbox used by Firefox here).  In the meantime, add libpciaccess explicitly
to fix this issue.